Add check_for_duplicates validator
